Library_Management/requirements.txt
streamlit
Pillow # Tkinter requires Pillow for some image handling, if needed
